requirements.txt
django
pillow
django-cleanup
django-allauth
django-htmx
